Added script wp v2

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -22,7 +22,7 @@ void print_number(int n)
 
 	while (bool)
 	{
-		if (num / ((long)counter * 10) > 0)
+		if (num / (counter * 10) > 0)
 			counter *= 10;
 		else
 		bool = 0;
